feat: add function to handle slugs
