--- src/utils/generate-slug.ts
@@ -1,9 +1,0 @@
-export function generateSlug(title: string): string {
-    return title
-    .toLowerCase()
-    .normalize("NFD").replace(/[\u0300-\u036f]/g, "")
-    .replace(/\//g, " ")
-    .replace(/[^a-z0-9 ]/g, "")
-    .trim()
-    .replace(/\s+/g, "-");
-}
